refactor: move inline model classes to JIM.Models and add model placement rule (#394)
Move 5 model/POCO classes from JIM.Application service files to their
own files in JIM.Models, enforcing the one-class-per-file convention:

- PendingExportReconciliationResult → JIM.Models/Transactional/
- FileSystemListResult, FileSystemEntry → JIM.Models/Utility/
- LogFileInfo, LogEntry → JIM.Models/Utility/

Add CLAUDE.md rule: all model classes must live in JIM.Models, never
inline in service files. Exceptions: JIM.Web/Models/ for UI-specific
models and JIM.Web/Models/Api/ for API DTOs.
