+++ src/JIM.Web/Controllers/Api/LogsController.cs
@@ -1,5 +1,6 @@
 using Asp.Versioning;
 using JIM.Application.Services;
+using JIM.Models.Utility;
 using JIM.Utilities;
 using JIM.Web.Models.Api;
 using Microsoft.AspNetCore.Authorization;
