+++ src/JIM.Web/Pages/Admin/Logs.razor
@@ -1,6 +1,7 @@
 @page "/admin/logs"
 @attribute [Authorize(Roles = "Administrator")]
 @using JIM.Application.Services
+@using JIM.Models.Utility
 @using JIM.Web.Models.Api
 @using JIM.Web.Services
 @using System.Text.Json
