+++ src/JIM.Web/Shared/FileBrowserDialog.razor
@@ -1,4 +1,5 @@
 @using JIM.Application.Servers
+@using JIM.Models.Utility
 @inject IJimApplicationFactory JimFactory
 
 <MudDialog>
