+++ src/JIM.Worker/Processors/SyncImportTaskProcessor.cs
@@ -10,6 +10,7 @@ using JIM.Models.Exceptions;
 using JIM.Models.Interfaces;
 using JIM.Models.Staging;
 using JIM.Models.Tasking;
+using JIM.Models.Transactional;
 using Serilog;
 using JIM.Worker.Models;
